+++ .github/workflows/compile_nuttx.yml
@@ -39,7 +39,7 @@ jobs:
           nxp_fmuk66-v3_socketcan,
           nxp_fmuk66-v3_rtps,
           nxp_fmurt1062-v1_default,
-          nxp_rddrone-uavcan146_default,
+          nxp_ucans32k146_default,
           omnibus_f4sd_default,
           px4_fmu-v2_default,
           px4_fmu-v2_fixedwing,
