+++ boards/nxp/ucans32k146/firmware.prototype
@@ -1,10 +1,10 @@
 {
     "board_id": 10000,
     "magic": "PX4FWv1",
-    "description": "Firmware for the rddrone-uavcan146 board",
+    "description": "Firmware for the ucans32k146 board",
     "image": "",
     "build_time": 0,
-    "summary": "RDDRONE-UAVCAN146",
+    "summary": "UCANS32K146",
     "version": "0.1",
     "image_size": 0,
     "image_maxsize": 0,
