+++ boards/nxp/ucans32k146/init/rc.board_defaults
