+++ boards/nxp/ucans32k146/nuttx-config/Kconfig
