+++ boards/nxp/ucans32k146/nuttx-config/scripts/s32k146.cfg
