+++ boards/nxp/ucans32k146/nuttx-config/scripts/sram.ld
@@ -1,5 +1,5 @@
 /****************************************************************************
- * boards/arm/s32k1xx/rddrone-uavcan146/scripts/sram.ld
+ * boards/arm/s32k1xx/ucans32k146/scripts/sram.ld
  *
  *   Copyright (C) 2019 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
