+++ boards/nxp/ucans32k146/src/periphclocks.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * boards/arm/s32k1xx/rddrone-uavcan146/src/s32k1xx_periphclks.c
+ * boards/arm/s32k1xx/ucans32k146/src/s32k1xx_periphclks.c
  *
  *   Copyright (C) 2019 Gregory Nutt. All rights reserved.
  *   Author: Gregory Nutt <gnutt@nuttx.org>
